Add dynamic user and loket info to SQL queries
Added a new `placeholders` parameter to `PaketCommand.cs` to include a placeholder for `[bsbs]` with the value `AppSettings.DBNameBilling`.

In `rekening_air_angsuran.sql`, created a temporary table `__tmp_userloket` to store user information and used it to replace the hardcoded `0` value for `iduser` with `usr.iduser`.

In `rekening_air_angsuran_detail.sql`, replaced `temp_dataawal_periode` with `__tmp_periode`, and created new temporary tables `__tmp_userloket` and `__tmp_loket` to store user and loket information. These tables are used to replace `NULL` values for `iduser` and `idloket` with `usr.iduser` and `lo.idloket`, respectively.

These changes ensure that user and loket information is dynamically retrieved and used in the queries, improving the accuracy and maintainability of the code.